Refactor storage and group common test code in specific file
